+++ chen_cppcguilib/basicProgressBar.h
@@ -5,8 +5,8 @@ class basicProgressBar : public component
 {
 public:
 	basicProgressBar(int length, int style);
-	int getWidth() const override;
-	int getHeight() const override;
+	size_t getWidth() const override;
+	size_t getHeight() const override;
 	std::vector<cgui::string> getData() const override;
 
 	void updateProgress(int progress);
