+++ chen_cppcguilib/basicText.h
@@ -5,8 +5,8 @@ class basicText : public component
 {
 public:
     basicText(std::string_view str);
-    int getWidth() const override;
-    int getHeight() const override;
+    size_t getWidth() const override;
+    size_t getHeight() const override;
     std::vector<cgui::string> getData() const override;
     
     void setText(std::string_view str);
