+++ chen_cppcguilib/component.h
@@ -11,7 +11,7 @@
 class component
 {
 public:
-    virtual int getWidth() const = 0;
-    virtual int getHeight() const = 0;
+    virtual size_t getWidth() const = 0;
+    virtual size_t getHeight() const = 0;
     virtual std::vector<cgui::string> getData() const = 0;
 };
