+++ chen_cppcguilib/multiLineText.h
@@ -6,13 +6,13 @@ class multiLineText : public component
 {
 public:
     multiLineText(std::vector<cgui::string> textByLine);
-    int getWidth() const override;
-    int getHeight() const override;
+    size_t getWidth() const override;
+    size_t getHeight() const override;
     std::vector<cgui::string> getData() const override;
     void setMultiText(std::vector<cgui::string> textByLine);
 
 private:
     std::vector<cgui::string> textByLine;
-    int calcuateLength() const;
+    size_t calcuateLength() const;
 };
