+++ chen_cppcguilib/utils/string.h
@@ -26,7 +26,7 @@ public:
     string(int count, char c);
 
     size_t size() const;
-    int length() const;
+    size_t length() const;
     
     void insert(int pos, const string& other);
     void insert(int pos, int count, char c);
@@ -49,7 +49,7 @@ public:
 
 private:
     std::string str;
-    int visibleLength = 0;
+    size_t visibleLength = 0;
     int colorCount = 0; // 用于颜色的ascii转义字符数量
 
     int pushBackPos() const; // 如果有颜色，pos会在最后一个恢复默认颜色之前
